+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -50,7 +50,7 @@
   - title: Dialog
     url: "/components/alpha/dialog"
   - title: Dropdown
-    url: "/components/dropdown"
+    url: "/components/alpha/dropdown"
   - title: DropdownMenu
     url: "/components/dropdownmenu"
   - title: Flash
