+++ previews/primer/alpha/dropdown_preview.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 module Primer
+  module Alpha
   # @label Dropdown
   class DropdownPreview < ViewComponent::Preview
     # @label Playground
@@ -8,7 +9,7 @@ module Primer
     # @param with_caret [Boolean] toggle
     # @param overlay [Symbol] select [none, default, dark]
     def playground(overlay: :default, with_caret: false)
-      render(Primer::Dropdown.new(overlay: overlay, with_caret: with_caret)) do |c|
+      render(Primer::Alpha::Dropdown.new(overlay: overlay, with_caret: with_caret)) do |c|
         c.with_button { "Dropdown" }
 
         c.with_menu(header: "Header") do |m|
@@ -24,7 +25,7 @@ module Primer
     # @label Default
     #
     def default
-      render(Primer::Dropdown.new) do |c|
+      render(Primer::Alpha::Dropdown.new) do |c|
         c.with_button { "Dropdown" }
         c.with_menu do |m|
           m.with_item { "Item 1" }
@@ -40,7 +41,7 @@ module Primer
     # @param direction [Symbol] select [se, sw, w, e, ne, s]
     # @param scheme [Symbol] select [default, dark]
     def menu(as: :default, direction: :se, scheme: :default)
-      render(Primer::Dropdown::Menu.new(as: as, direction: direction, scheme: scheme, header: "Header")) do |m|
+      render(Primer::Alpha::Dropdown::Menu.new(as: as, direction: direction, scheme: scheme, header: "Header")) do |m|
         m.with_item { "Item 1" }
         m.with_item { "Item 2" }
         m.with_item(divider: true)
@@ -53,7 +54,7 @@ module Primer
     #
     # @label Direction e
     def direction_e
-      render(Primer::Dropdown.new(display: :inline_block)) do |c|
+      render(Primer::Alpha::Dropdown.new(display: :inline_block)) do |c|
         c.with_button { "Dropdown" }
         c.with_menu(direction: :e) do |m|
           m.with_item { "Item 1" }
@@ -65,7 +66,7 @@ module Primer
 
     # @label Direction ne
     def direction_ne
-      render(Primer::Dropdown.new(display: :inline_block)) do |c|
+      render(Primer::Alpha::Dropdown.new(display: :inline_block)) do |c|
         c.with_button { "Dropdown" }
         c.with_menu(direction: :ne) do |m|
           m.with_item { "Item 1" }
@@ -77,7 +78,7 @@ module Primer
 
     # @label Direction s
     def direction_s
-      render(Primer::Dropdown.new(display: :inline_block)) do |c|
+      render(Primer::Alpha::Dropdown.new(display: :inline_block)) do |c|
         c.with_button { "Dropdown" }
         c.with_menu(direction: :s) do |m|
           m.with_item { "Item 1" }
@@ -89,7 +90,7 @@ module Primer
 
     # @label Direction se
     def direction_se
-      render(Primer::Dropdown.new(display: :inline_block)) do |c|
+      render(Primer::Alpha::Dropdown.new(display: :inline_block)) do |c|
         c.with_button { "Dropdown" }
         c.with_menu(direction: :se) do |m|
           m.with_item { "Item 1" }
@@ -101,7 +102,7 @@ module Primer
 
     # @label Direction sw
     def direction_sw
-      render(Primer::Dropdown.new(display: :inline_block)) do |c|
+      render(Primer::Alpha::Dropdown.new(display: :inline_block)) do |c|
         c.with_button { "Dropdown" }
         c.with_menu(direction: :sw) do |m|
           m.with_item { "Item 1" }
@@ -113,7 +114,7 @@ module Primer
 
     # @label Direction w
     def direction_w
-      render(Primer::Dropdown.new(display: :inline_block)) do |c|
+      render(Primer::Alpha::Dropdown.new(display: :inline_block)) do |c|
         c.with_button { "Dropdown" }
         c.with_menu(direction: :w) do |m|
           m.with_item { "Item 1" }
@@ -129,7 +130,7 @@ module Primer
     #
     # @label With caret
     def options_with_caret
-      render(Primer::Dropdown.new(with_caret: true)) do |c|
+      render(Primer::Alpha::Dropdown.new(with_caret: true)) do |c|
         c.with_button { "Dropdown" }
         c.with_menu do |m|
           m.with_item { "Item 1" }
@@ -141,7 +142,7 @@ module Primer
 
     # @label With header
     def options_with_header
-      render(Primer::Dropdown.new) do |c|
+      render(Primer::Alpha::Dropdown.new) do |c|
         c.with_button { "Dropdown" }
         c.with_menu(header: "Header") do |m|
           m.with_item { "Item 1" }
@@ -153,7 +154,7 @@ module Primer
 
     # @label With dividers
     def options_with_dividers
-      render(Primer::Dropdown.new) do |c|
+      render(Primer::Alpha::Dropdown.new) do |c|
         c.with_button { "Dropdown" }
         c.with_menu do |m|
           m.with_item { "Item 1" }
@@ -169,7 +170,7 @@ module Primer
 
     # @label As list
     def options_as_list
-      render(Primer::Dropdown.new) do |c|
+      render(Primer::Alpha::Dropdown.new) do |c|
         c.with_button { "Dropdown" }
         c.with_menu(as: :list) do |m|
           m.with_item { "Item 1" }
@@ -181,7 +182,7 @@ module Primer
 
     # @label Overlay none
     def options_overlay_none
-      render(Primer::Dropdown.new(overlay: :none)) do |c|
+      render(Primer::Alpha::Dropdown.new(overlay: :none)) do |c|
         c.with_button { "Dropdown" }
         c.with_menu do |m|
           m.with_item { "Item 1" }
@@ -193,7 +194,7 @@ module Primer
 
     # @label Overlay dark
     def options_overlay_dark
-      render(Primer::Dropdown.new(overlay: :dark)) do |c|
+      render(Primer::Alpha::Dropdown.new(overlay: :dark)) do |c|
         c.with_button { "Dropdown" }
         c.with_menu do |m|
           m.with_item { "Item 1" }
@@ -204,5 +205,6 @@ module Primer
     end
     #
     # @!endgroup
+  end
   end
 end
