+++ static/audited_at.json
@@ -11,6 +11,9 @@
   "Primer::Alpha::Dialog::Body": "",
   "Primer::Alpha::Dialog::Footer": "",
   "Primer::Alpha::Dialog::Header": "",
+  "Primer::Alpha::Dropdown": "",
+  "Primer::Alpha::Dropdown::Menu": "",
+  "Primer::Alpha::Dropdown::Menu::Item": "",
   "Primer::Alpha::HiddenTextExpander": "",
   "Primer::Alpha::Image": "",
   "Primer::Alpha::ImageCrop": "",
@@ -63,7 +66,6 @@
   "Primer::Content": "",
   "Primer::Dropdown": "",
   "Primer::Dropdown::Menu": "",
-  "Primer::Dropdown::Menu::Item": "",
   "Primer::DropdownMenuComponent": "",
   "Primer::HellipButton": "",
   "Primer::IconButton": "",
