+++ static/statuses.json
@@ -11,6 +11,9 @@
   "Primer::Alpha::Dialog::Body": "alpha",
   "Primer::Alpha::Dialog::Footer": "alpha",
   "Primer::Alpha::Dialog::Header": "alpha",
+  "Primer::Alpha::Dropdown": "alpha",
+  "Primer::Alpha::Dropdown::Menu": "alpha",
+  "Primer::Alpha::Dropdown::Menu::Item": "alpha",
   "Primer::Alpha::HiddenTextExpander": "alpha",
   "Primer::Alpha::Image": "alpha",
   "Primer::Alpha::ImageCrop": "alpha",
@@ -61,9 +64,8 @@
   "Primer::ClipboardCopy": "beta",
   "Primer::ConditionalWrapper": "alpha",
   "Primer::Content": "stable",
-  "Primer::Dropdown": "alpha",
-  "Primer::Dropdown::Menu": "alpha",
-  "Primer::Dropdown::Menu::Item": "alpha",
+  "Primer::Dropdown": "deprecated",
+  "Primer::Dropdown::Menu": "deprecated",
   "Primer::DropdownMenuComponent": "deprecated",
   "Primer::HellipButton": "alpha",
   "Primer::IconButton": "deprecated",
