+++ test/components/alpha/dropdown/menu_test.rb
@@ -2,37 +2,37 @@
 
 require "components/test_helper"
 
-class Primer::Dropdown::MenuTest < Minitest::Test
+class Primer::Alpha::Dropdown::MenuTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_dropdown_component_renders_dark_scheme
-    render_inline(Primer::Dropdown::Menu.new(scheme: :dark)) { "Body" }
+    render_inline(Primer::Alpha::Dropdown::Menu.new(scheme: :dark)) { "Body" }
 
     assert_selector("details-menu.dropdown-menu.dropdown-menu-dark")
   end
 
   def test_dropdown_direction_renders
-    render_inline(Primer::Dropdown::Menu.new(direction: :w)) { "Body" }
+    render_inline(Primer::Alpha::Dropdown::Menu.new(direction: :w)) { "Body" }
 
     assert_selector("details-menu.dropdown-menu.dropdown-menu-w")
   end
 
   def test_falls_back_to_defaults_when_invalid_options_are_passed
     without_fetch_or_fallback_raises do
-      render_inline(Primer::Dropdown::Menu.new(direction: :circle, scheme: :orange)) { "Body" }
+      render_inline(Primer::Alpha::Dropdown::Menu.new(direction: :circle, scheme: :orange)) { "Body" }
     end
 
     assert_selector("details-menu.dropdown-menu")
   end
 
   def test_dropdown_header_renders
-    render_inline(Primer::Dropdown::Menu.new(header: "Header")) { "Body" }
+    render_inline(Primer::Alpha::Dropdown::Menu.new(header: "Header")) { "Body" }
 
     assert_selector(".dropdown-header")
   end
 
   def test_renders_items
-    render_inline(Primer::Dropdown::Menu.new(header: "Header")) do |c|
+    render_inline(Primer::Alpha::Dropdown::Menu.new(header: "Header")) do |c|
       c.item { "Item 1" }
       c.item { "Item 2" }
     end
@@ -42,7 +42,7 @@ class Primer::Dropdown::MenuTest < Minitest::Test
   end
 
   def test_renders_items_as_buttons
-    render_inline(Primer::Dropdown::Menu.new(header: "Header")) do |c|
+    render_inline(Primer::Alpha::Dropdown::Menu.new(header: "Header")) do |c|
       c.item(tag: :button) { "Item 1" }
       c.item(tag: :button) { "Item 2" }
     end
@@ -52,7 +52,7 @@ class Primer::Dropdown::MenuTest < Minitest::Test
   end
 
   def test_renders_dividers
-    render_inline(Primer::Dropdown::Menu.new(header: "Header")) do |c|
+    render_inline(Primer::Alpha::Dropdown::Menu.new(header: "Header")) do |c|
       c.item(divider: true)
     end
 
@@ -60,7 +60,7 @@ class Primer::Dropdown::MenuTest < Minitest::Test
   end
 
   def test_renders_as_list
-    render_inline(Primer::Dropdown::Menu.new(as: :list, header: "Header")) do |c|
+    render_inline(Primer::Alpha::Dropdown::Menu.new(as: :list, header: "Header")) do |c|
       c.item { "Item 1" }
       c.item(divider: true)
       c.item { "Item 2" }
