+++ test/css/component_selector_use_test.rb
@@ -8,7 +8,7 @@ IGNORED_SELECTORS = {
   :global => ["preview-wrap"],
   Primer::Alpha::TabPanels => ["tabnav", "tabnav-tab", "tabnav-tabs"],
   Primer::Beta::Button => ["Button--medium", "octicon", "octicon-search", "Button--invisible-noVisuals"],
-  Primer::Dropdown => ["btn", "details-overlay", "details-reset", "dropdown-menu-se", "octicon", "octicon-triangle-down"],
+  Primer::Alpha::Dropdown => ["btn", "details-overlay", "details-reset", "dropdown-menu-se", "octicon", "octicon-triangle-down"],
   Primer::Beta::AvatarStack => ["AvatarStack", "AvatarStack-body", "avatar", "avatar-small", "tooltipped", "tooltipped-n"],
   Primer::Beta::CloseButton => ["close-button", "octicon", "octicon-x"],
   Primer::Alpha::TextField => ["FormControl", "FormControl-input", "FormControl-label"],
